Fix path + naming
Right now the generator is adding a "/" in front of reference/recipes
for all of the generated recipes. GitBook is removing that which
means every time we re-generate docs, we get a huge changelog of
unnecessary and confusing changes. This should fix that.

Also ensures that "`" do not appear in display names for categories.