Use versionless Maven commands
Fixes https://github.com/openrewrite/rewrite-docs/issues/120

Has a slight potential to make it harder to debug when a user reports a bug in say january, and we don't immediately know what version they used. But the versions will still be in the logs, and the convenience might out weigh the negative.