Fix recipe link urls
We previously extracted the recipes from reference/recipes to just recipes in rewrite-docs. That being said, that was just visually. Behind the scenes, it's still stored in /reference/recipes. However, GitBook decided to not reference the docs from the path but rather from the display URL. This can sometimes result in doc links not working. This change should fix that