Drop contributors from individual recipe pages (#220)
* Drop contributors from individual recipe pages

We had some concerns reported to us that not every contributor has signed up to be listed here, so to avoid issues we're for now removing their names here. The metadata is retained, and we continue to appreciate the work done previously and going forward!

* Drop contributors from individual recipe pages

We had some concerns reported to us that not every contributor has signed up to be listed here, so to avoid issues we're for now removing their names here. The metadata is retained, and we continue to appreciate the work done previously and going forward!