Add support for JavaScript docs (#240)
There are some caveats here - type functionality doesn't work as the RPC
doesn't support that (a known bug). Also, some JS recipes haven't been
published - such as the react ones, so we can't include them in the docs.

This is a first draft to just have some details for JS recipes but more
work/changes will need to be done in the future as we implement things like
running JS recipes in app.moderne or whatnot.

Co-authored-by: team-moderne[bot] <mike-solomon@users.noreply.github.com>